melee, range and static complete
