+++ src/aa-classes/animation-data.js
@@ -1,4 +1,5 @@
 import { buildFile } from "../animation-functions/file-builder/build-filepath.js";
+import { particleDefaultValues } from "../animation-functions/particleDefaults.js";
 
 export class AAanimationData {
 
@@ -323,4 +324,78 @@ export class AAanimationData {
                 break
         }
     }
+
+    static async compileParticleData(autoObject, options3d) {
+        const type = options3d.type;
+        if (autoObject) {
+            const explode3d = options3d.addexplosion || {};
+            const data = {
+                type: type,
+                speed: options3d.speed ?? particleDefaultValues[type].speed,
+                repeat: options3d.repeat ?? particleDefaultValues[type].repeat,
+                delay: options3d.delay ?? particleDefaultValues[type].delay,
+                arc: options3d.arc ?? particleDefaultValues[type].arc,
+                scale: options3d.scale ?? particleDefaultValues[type].scale,
+                color01: options3d.color01 ?? particleDefaultValues[type].color01,
+                color02: options3d.color02 ?? particleDefaultValues[type].color02,
+                sprite: options3d.sprite ?? particleDefaultValues[type].sprite,
+                life: options3d.life ?? particleDefaultValues[type].life,
+                emittersize: options3d.emittersize ?? particleDefaultValues[type].emittersize,
+                alpha: options3d.alpha ?? particleDefaultValues[type].alpha,
+                mass: options3d.mass ?? particleDefaultValues[type].mass,
+                gravity: options3d.gravity ?? particleDefaultValues[type].gravity,
+                rate: options3d.rate ?? particleDefaultValues[type].rate,
+                explosion: {
+                    enable: explode3d.enable || false,
+                    color01: explode3d.color01 ?? particleDefaultValues[type].explosion.color01,
+                    color02: explode3d.color02 ?? particleDefaultValues[type].explosion.color02,
+                    speed: explode3d.speed ?? particleDefaultValues[type].explosion.speed,
+                    gravity: explode3d.gravity ?? particleDefaultValues[type].explosion.gravity,
+                    life: explode3d.life ?? particleDefaultValues[type].explosion.life,
+                    rate: explode3d.rate ?? particleDefaultValues[type].explosion.rate,
+                    emittersize: explode3d.emittersize ?? particleDefaultValues[type].explosion.emittersize,
+                    alpha: explode3d.alpha ?? particleDefaultValues[type].explosion.alpha,
+                    mass: explode3d.mass ?? particleDefaultValues[type].explosion.mass,
+                    sprite: explode3d.sprite ?? particleDefaultValues[type].explosion.sprite,
+                    scale: explode3d.scale ?? particleDefaultValues[type].explosion.scale,
+                }
+            };
+            return data;
+        } else {
+            const explode3d = options3d[type].addexplosion || {};
+            const data = {
+                type: type,
+                speed: options3d[type].speed ?? particleDefaultValues[type].speed,
+                repeat: options3d[type].repeat ?? particleDefaultValues[type].repeat,
+                delay: options3d[type].delay ?? particleDefaultValues[type].delay,
+                arc: options3d[type].arc ?? particleDefaultValues[type].arc,
+                scale: options3d[type].scale ?? particleDefaultValues[type].scale,
+                color01: options3d[type].color01 ?? particleDefaultValues[type].color01,
+                color02: options3d[type].color02 ?? particleDefaultValues[type].color02,
+                sprite: options3d[type].sprite ?? particleDefaultValues[type].sprite,
+                life: options3d[type].life ?? particleDefaultValues[type].life,
+                emittersize: options3d[type].emittersize ?? particleDefaultValues[type].emittersize,
+                alpha: options3d[type].alpha ?? particleDefaultValues[type].alpha,
+                mass: options3d[type].mass ?? particleDefaultValues[type].mass,
+                gravity: options3d[type].gravity ?? particleDefaultValues[type].gravity,
+                rate: options3d[type].rate ?? particleDefaultValues[type].rate,
+                explosion: {
+                    enable: explode3d.enable || false,
+                    color01: explode3d.color01 ?? particleDefaultValues[type].explosion.color01,
+                    color02: explode3d.color02 ?? particleDefaultValues[type].explosion.color02,
+                    speed: explode3d.speed ?? particleDefaultValues[type].explosion.speed,
+                    gravity: explode3d.gravity ?? particleDefaultValues[type].explosion.gravity,
+                    life: explode3d.life ?? particleDefaultValues[type].explosion.life,
+                    rate: explode3d.rate ?? particleDefaultValues[type].explosion.rate,
+                    emittersize: explode3d.emittersize ?? particleDefaultValues[type].explosion.emittersize,
+                    alpha: explode3d.alpha ?? particleDefaultValues[type].explosion.alpha,
+                    mass: explode3d.mass ?? particleDefaultValues[type].explosion.mass,
+                    sprite: explode3d.sprite ?? particleDefaultValues[type].explosion.sprite,
+                    scale: explode3d.scale ?? particleDefaultValues[type].explosion.scale,
+                }
+            };
+            return data;
+        }
+    }
+
 }
